\name{logit}
\title{Logistic transform}
\alias{logit}
\alias{inv.logit}
\description{
Logistic and inverse logistic transform. For internal use.
}
\usage{
logit(x) 

inv.logit(x) 
}
\arguments{
  \item{x}{a \emph{numeric vector}, ranging between 0 and 1 for \code{logit}.}  
}
\value{
  A \emph{numeric vector}.
}
\keyword{functions,internal}
