\docType{package}
\name{MRIaggr-package}
\title{
Management, display and processing of cerebral imaging data.
}
\alias{MRIaggr-package}
\description{
MRIaggr provides a compact storage for medical imaging data with access and display possibilities. 
Basic tools for processing brain imaging data are proposed like extraction of brain voxels, morphological image segmentation, and filtering / normalization of contrast parameters. 
Specific tools are also provided for blood perfusion imaging to calculate hypoperfusion and reperfusion volumes.
}
\details{
\tabular{ll}{
Package: \tab MRIaggr\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-06-25\cr
License: GPL ( >=2.0) \tab \cr
}
}
\author{
brice ozenne

Maintainer: brice ozenne <brice.ozenne@orange.fr>
}
\keyword{package}


