\name{MRIaggr.Pat1_red}
\docType{data}
\alias{MRIaggr.Pat1_red}
\title{Example of processed MRIaggr object}
\description{
Example of processed \code{\linkS4class{MRIaggr}} object using the functionalities of the MRIaggr package.
}
\usage{MRIaggr.Pat1_red}
\format{A \code{\linkS4class{MRIaggr}} object containing in addition to the contrast data :
\itemize{
\item the position of the mid-saggital plan.
\item the normalization values for the contrast parameters.
\item the spatial groups of lesion.
\item the regional and controlateral normalised contrast values.
\item the  CSF/WM/GM probabilistic membership.
\item the neighborhood matrix.
\item the lesion volumes per slice
\item the hypoperfusion and reperfusion volumes for various time thresholds. 
}
}
\source{I-know study : \url{www.i-know-stroke.eu}}
\keyword{datasets}