\name{allocOptionsMRIaggr}
\title{Allocate new default values}
\alias{allocOptionsMRIaggr}
\description{
Allocate default value(s) to \code{\link{optionsMRIaggr}}. For internal use.
}
\usage{
allocOptionsMRIaggr(field, n.args)
}
\arguments{
  \item{field}{a named list containing the value(s) to be allocated. \emph{list}}
  \item{n.args}{the number of default value(s) to be allocated. \emph{positive integer}}
}
\concept{affect.}
\keyword{function,internal}
