\name{selectOptionsMRIaggr}
\title{Extract default values}
\alias{selectOptionsMRIaggr}
\description{
Extract value(s) from \code{\link{optionsMRIaggr}}. For internal use.
}
\usage{
selectOptionsMRIaggr(field = NULL)
}
\arguments{
  \item{field}{the field to be extracted. \emph{character}}
}
\concept{select.}
\keyword{function,internal}
